Cleanup Encoding tests
- GetEncoding
- Convert
